services/model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "YOLO11 Model Conversion Tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "textual>=6.3.0",
    "ultralytics>=8.3.217",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "torch>=2.5.0",
    "pillow>=11.0.0",
    "opencv-python>=4.10.0",
    "numpy>=2.1.0",
]

[project.scripts]
model-convertor = "model-convertor.main:main"

[tool.poetry]
name = "model"
version = "0.1.0"
description = "YOLO11 Model Conversion Tool and training script"
authors = [
    { name = "qwertuhh" }
]
